Reorganize oauth strategy files
