+++ lib/civicrm/api.rb
@@ -1,0 +1,4 @@
+require "civicrm/api/address"
+require "civicrm/api/error"
+require "civicrm/api/request"
+require "civicrm/api/user"
