+++ lib/civicrm/api/error.rb
@@ -1,0 +1,11 @@
+module Civicrm
+  module Api
+    class Error < StandardError
+      def initialize(detail = nil)
+        msg = "There was an error with the CiViCRM API"
+        msg += " (#{detail})" if detail.present?
+        super(msg)
+      end
+    end
+  end
+end
