--- lib/civicrm_api/civicrm_api.rb
@@ -1,3 +1,0 @@
-require "civicrm_api/request"
-require "civicrm_api/models/user"
-require "civicrm_api/models/address"
